sqlite and api

+++ Data/Db.cs
@@ -1,19 +1,48 @@
 ﻿using Restaurant_Manager.Model;
 using Restaurant_Manager.Model.Enums;
 using SQLite;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Restaurant_Manager.Data
 {
+    /// <summary>
+    /// THIS IS THE MAIN DATABASE FILE 
+    /// IMPORTANT(((IF YOU ARE WORKING ON UI JUST USE THE DatabaseApi.cs class i made that so you can call methods EASIER AND FASTER WITHOUT READING THIS!)))
+    /// THIS CLASS INITIALIZES SQLite and creates the .db file
+    /// HAS CRUD METHODS FOR ALL Model objects in the program
+    /// CHECKS FOR CONFLICTS AND CREATES INDEXES in case of RESERVATIONS/TABLES
+    /// AT THE MOMENT IT ALSO DELETES AND RE-CREATES DATABASE ON EACH RUN SO ITS EASIER FOR DEBUGGING
+    /// AND ALSO DOES SEEDING AFTER CREATING THE DATABASE
+    /// YOUSSIF AL-HALAWCHE
+    /// </summary>
     public static class Db
     {
+        //RESET DATABASE OF EVERY RUN FLAG; USE THIS WHEN YOU WANT TO START WITH A FRESH DATABASE 
+        //
+        public static bool ResetDbOnStart { get; set; } = false;
+        //The Connection object and the initialized boolean are global so methods are able to check of connection/initialization status
         static SQLiteAsyncConnection? _conn;
         static bool _initialized;
 
+        //MENU ITEMS AND EMPLOYEES SEEDING LISTS
+
+        static readonly MenuEntry[] menuEntries = new[]
+        {
+            new MenuEntry { Name="Margherita Pizza", Price=12.5m, IsAvailable=true, Type=MenuEntryType.MainDish },
+            new MenuEntry { Name="Shawrma Chicken Wrap", Price=6.0m, IsAvailable=true, Type=MenuEntryType.MainDish },
+            new MenuEntry { Name="Diet Coke", Price=2.0m, IsAvailable=true, Type=MenuEntryType.Drink },
+            new MenuEntry { Name="Chicken Biryani", Price=12.0m, IsAvailable=true, Type=MenuEntryType.MainDish },
+            new MenuEntry { Name="Chocolate Cake", Price=6.0m, IsAvailable=true, Type=MenuEntryType.Dessert },
+            new MenuEntry { Name="Lamb on rice", Price=17.0m, IsAvailable=true, Type=MenuEntryType.MainDish },
+        };
+        static readonly Employee[] EmployeesList = new[]
+        {
+        new Employee { Name="Someone", Age=26, City="Calgary", Country="CA" },
+        new Employee { Name="Youssif", Age=24, City="Calgary", Country="CA" },
+        new Employee { Name="Paulie", Age=67, City="New Jersey", Country="US" },
+        new Employee { Name="Tony", Age=44, City="New Jersey", Country="US" },
+        new Employee { Name="Aidriana", Age=24, City="Calgary", Country="CA" },
+        };
+        ///
         public static async Task Initialize()
         {
             //check if the databse is already initialzied, if so returns
@@ -23,25 +52,27 @@ namespace Restaurant_Manager.Data
             }
             //establish connection
             var path = Path.Combine(FileSystem.AppDataDirectory, "res.db");
+            System.Diagnostics.Debug.WriteLine($"======================== PATH IS AT {path} ========================");
             //Added this check to delete previous databases if found
             //so each run of the app starts with a fresh version of the database
             //to avoid confusion and cut on 
             //debugging time for this assignemnt
-            if(File.Exists(path))
+            if(ResetDbOnStart && File.Exists(path))
             {
                 File.Delete(path);
             }
             _conn = new SQLiteAsyncConnection(path);
             //and create tables for our models
+            //
             await _conn.CreateTableAsync<Employee>();
             await _conn.CreateTableAsync<MenuEntry>();
             await _conn.CreateTableAsync<Table>();
             await _conn.CreateTableAsync<Reservation>();
-
+            //Create indexes for tables for faster lookups
             await _conn.CreateIndexAsync(nameof(Table), new[] { nameof(Table.TableNumber) }, unique: true);
             await _conn.CreateIndexAsync(nameof(Reservation),
                 new[] { nameof(Reservation.TableId), nameof(Reservation.StartUtc), nameof(Reservation.EndUtc) }, unique: false);
-
+             
             if (await _conn.Table<Table>().CountAsync() == 0)
             {
                 await _conn.InsertAllAsync(new[]
@@ -58,34 +89,12 @@ namespace Restaurant_Manager.Data
             //
             if (await _conn.Table<MenuEntry>().CountAsync() == 0)
             {
-                await _conn.InsertAllAsync(new[]
-                {
-        new MenuEntry { Name="Margherita Pizza", Price=12.5m, IsAvailable=true, Type=MenuEntryType.MainDish },
-        new MenuEntry { Name="Shawrma Chicken Wrap",   Price=6.0m,  IsAvailable=true, Type=MenuEntryType.MainDish  },
-        new MenuEntry { Name="Diet Coke",   Price=2.0m,  IsAvailable=true, Type=MenuEntryType.Drink},
-        new MenuEntry { Name="Chicken Biryani",   Price=12.0m,  IsAvailable=true, Type=MenuEntryType.MainDish  },
-        new MenuEntry { Name="Chocolate Cake",   Price=6.0m,  IsAvailable=true, Type=MenuEntryType.Dessert  },
-        new MenuEntry { Name="Lamb on rice",   Price=17.0m,  IsAvailable=true, Type=MenuEntryType.MainDish  },
-
-
-
-
-    });
+                await _conn.InsertAllAsync(menuEntries);
             }
             //If we do not have employees then seed the table with employees same procedure as the previous seeding of MenuItems
             if (await _conn.Table<Employee>().CountAsync() == 0)
             {
-                await _conn.InsertAllAsync(new[]
-                {
-        new Employee { Name="Someone", Age=26, City="Calgary", Country="CA" },
-        new Employee { Name="Youssif", Age=24, City="Calgary", Country="CA" },
-        new Employee { Name="Paulie", Age=67, City="New Jersey", Country="US" },
-        new Employee { Name="Tony", Age=44, City="New Jersey", Country="US" },
-        new Employee { Name="Aidriana", Age=24, City="Calgary", Country="CA" },
-
-
-
-    });
+                await _conn.InsertAllAsync(EmployeesList);
             }
             //signal teh database initialized
             _initialized = true;
@@ -205,11 +214,16 @@ namespace Restaurant_Manager.Data
             return tables.Where(t => !busy.Contains(t.Id)).OrderBy(t => t.TableNumber).ToList();
         }
 
-        // ---------- Employees (basic) ----------
+        // ---------- Employees  ----------
         public static async Task<List<Employee>> GetEmployeesAsync()
         { await Initialize(); return await _conn!.Table<Employee>().OrderBy(e => e.Name).ToListAsync(); }
 
         public static async Task<int> SaveEmployeeAsync(Employee e)
         { await Initialize(); return e.Id == 0 ? await _conn!.InsertAsync(e) : await _conn!.UpdateAsync(e); }
+
+        public static async Task<int> DeleteEmployee(Employee e)
+        {
+            await Initialize(); return e.Id == 0 ? throw new InvalidOperationException("No employee found!") : await _conn!.DeleteAsync(e);
+        }
     }
 }
